feat: moved No Record Screen to separate folder
