--- app/cdap/components/GridTable/components/NoRecordScreen/constants.ts
@@ -1,2 +1,0 @@
-export const SUB_HEADER_TEXT = 'Remove some recipe steps to view some data';
-export const MAIN_HEADER_TEXT = 'No rows left in this sample';
