+++ app/cdap/components/NoRecordScreen/__tests__/NoRecordScreen.test.tsx
