+++ app/cdap/components/NoRecordScreen/iconStore.tsx
