+++ app/cdap/components/NoRecordScreen/styles.ts
